+++ src/main/java/dev/willyelton/crystal_tools/common/levelable/tool/CrystalTrident.java
@@ -4,6 +4,8 @@ package dev.willyelton.crystal_tools.common.levelable.tool;
 import dev.willyelton.crystal_tools.common.components.DataComponents;
 import dev.willyelton.crystal_tools.common.config.CrystalToolsConfig;
 import dev.willyelton.crystal_tools.common.entity.CrystalTridentEntity;
+import dev.willyelton.crystal_tools.common.events.LevelTickEvent;
+import dev.willyelton.crystal_tools.common.levelable.EntityTargeter;
 import net.minecraft.core.BlockPos;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.sounds.SoundEvents;
@@ -24,7 +26,7 @@ import net.minecraft.world.phys.Vec3;
 import net.neoforged.neoforge.common.ItemAbility;
 import org.jetbrains.annotations.NotNull;
 
-public class CrystalTrident extends SwordLevelableTool {
+public class CrystalTrident extends SwordLevelableTool implements EntityTargeter {
     public CrystalTrident() {
         super("trident", 4, -2.9F);
     }
@@ -61,18 +63,19 @@ public class CrystalTrident extends SwordLevelableTool {
 
     @Override
     public InteractionResultHolder<ItemStack> use(Level level, Player player, InteractionHand hand) {
-        ItemStack itemstack = player.getItemInHand(hand);
+        ItemStack stack = player.getItemInHand(hand);
+        refreshTarget(stack, level, player);
         if (this.isDisabled()) {
-            itemstack.shrink(1);
+            stack.shrink(1);
         }
 
-        if (itemstack.getDamageValue() >= itemstack.getMaxDamage() - 1) {
-            return InteractionResultHolder.fail(itemstack);
-        } else if (riptideEnabled(itemstack) && !canRiptide(itemstack, player)) {
-            return InteractionResultHolder.fail(itemstack);
+        if (stack.getDamageValue() >= stack.getMaxDamage() - 1) {
+            return InteractionResultHolder.fail(stack);
+        } else if (riptideEnabled(stack) && !canRiptide(stack, player)) {
+            return InteractionResultHolder.fail(stack);
         } else {
             player.startUsingItem(hand);
-            return InteractionResultHolder.consume(itemstack);
+            return InteractionResultHolder.consume(stack);
         }
     }
 
@@ -94,6 +97,11 @@ public class CrystalTrident extends SwordLevelableTool {
                         }
 
                         level.addFreshEntity(tridentEntity);
+                        int target = stack.getOrDefault(DataComponents.ENTITY_TARGET, -1);
+                        if (target != -1) {
+                            LevelTickEvent.startTracking(level, tridentEntity.getId(), target, velocity / 2);
+                        }
+
                         level.playSound(null, tridentEntity, SoundEvents.TRIDENT_THROW.value(), SoundSource.PLAYERS, 1.0F, 1.0F);
                         if (!player.getAbilities().instabuild) {
                             player.getInventory().removeItem(stack);
@@ -134,6 +142,19 @@ public class CrystalTrident extends SwordLevelableTool {
                 }
             }
         }
+    }
+
+    @Override
+    public void onUseTick(Level level, LivingEntity livingEntity, ItemStack stack, int remainingUseDuration) {
+        refreshTarget(stack, level, livingEntity);
+
+        super.onUseTick(level, livingEntity, stack, remainingUseDuration);
+    }
+
+    @Override
+    public void onStopUsing(ItemStack stack, LivingEntity entity, int count) {
+        clearTarget(stack, entity.level());
+        super.onStopUsing(stack, entity, count);
     }
 
     @Override